+++ src/test/php/lang/DefaultModeErrorHandlerTest.php
@@ -92,6 +92,7 @@ class ModeErrorHandler implements ErrorHandler
  *
  * @group  lang
  * @group  lang_core
+ * @deprecated  since 7.0.0, will be removed with 8.0.0
  */
 class ModeErrorHandlerTest extends \PHPUnit_Framework_TestCase
 {