+++ src/test/php/lang/DefaultModeTest.php
@@ -16,6 +16,7 @@ use function bovigo\assert\assertTrue;
  * All tests that do not require restoring the error or exception handler.
  *
  * @group  lang
+ * @deprecated  since 7.0.0, will be removed with 8.0.0
  */
 class DefaultModeTest extends \PHPUnit_Framework_TestCase
 {